# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de boutique",
        "description": "Documentation de l'API REST pour la gestion de boutique (catégories, produits, acheteurs, achats). CCP 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/produits": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Lister tous les produits (accessible sans connexion)",
                "operationId": "8b52d87531fe38cafe96328a8f1dd064",
                "responses": {
                    "200": {
                        "description": "Liste des produits, avec leur catégorie"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un nouveau produit",
                "operationId": "334cf1dab5c82124f1c45a4555ed7795",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prix",
                                    "stock",
                                    "categorie_id"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "T-shirt Nike"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "T-shirt 100% coton"
                                    },
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé (réservé gestionnaire/admin)"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/produits/{id}": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Afficher le détail d'un produit (avec sa catégorie et ses acheteurs)",
                "operationId": "9c5d60584961429a9c2a243463183cec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du produit"
                    },
                    "404": {
                        "description": "Produit introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produits"
                ],
                "summary": "Modifier un produit",
                "operationId": "42c2896c9129639cbcb0094ebb631890",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prix",
                                    "stock",
                                    "categorie_id"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "T-shirt Nike Pro"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 34.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 80
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Version mise à jour"
                                    },
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit modifié"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Produit introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produits"
                ],
                "summary": "Supprimer un produit",
                "operationId": "f9e972f1ce50047b09fd6db1ee1ec82d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produit supprimé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Produit introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez le token obtenu via /login, sans le mot 'Bearer' devant (Swagger l'ajoute automatiquement).",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}